Moved nnls to lib because it's from eigen
