+++ src/VRM.h
@@ -14,7 +14,7 @@
 #ifdef PLUGIN_SIMPLEX
 #include "simplex.h"
 #endif
-#include "nnls.h"
+#include "eigen-nnls/nnls.h"
 
 #include <Eigen/Dense>
